requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
tensorflow==2.7.0
python-dotenv
gunicorn
protobuf==3.9.2
keras-tuner
tensorboard==2.10.1